refactor: reports
