+++ commscores/report.py
@@ -1,3 +1,5 @@
+import os
+import re
 from itertools import chain, combinations, permutations
 
 from numpy import array, sort, unique, where
@@ -5,7 +7,9 @@ from numpy.random import shuffle
 
 from .logger import logger
 from .scores.calculate_scores import calculate_scores
-from .utils import _get_media
+from .utils import _get_media, convert_to_int, remove_metadata
+
+package_dir = os.path.abspath(os.path.dirname(__file__))
 
 
 def report_generation(
@@ -158,3 +162,194 @@ def report_generation(
             print_progress,
         )
     return concat(series, axis=1).T, mets
+
+
+def html_report(df, mets, export_html_path="commscores_report.html", msdb_path=None):
+    import jinja2
+    from pandas import to_numeric
+
+    def names_updateCPD(metIDs, update_cpdNames):
+        names = []
+        for metID in metIDs:
+            if metID not in cpdNames:
+                if "msdb" not in locals().keys():
+                    from modelseedpy.biochem import from_local
+
+                    msdb = from_local(msdb_path)
+                name = msdb.compounds.get(metID, None)
+                if name is None:
+                    name = metID
+                else:
+                    name = name["name"]
+                update_cpdNames[metID] = name
+            else:
+                name = cpdNames[metID]
+            names.append(name)
+        return names, update_cpdNames
+
+    # construct a heatmap
+    df.index.name = "Community_index"
+    heatmap_df = df.copy(deep=True)  # takes some time
+    heatmap_df_index = zip(
+        heatmap_df["model1"].to_numpy(), heatmap_df["model2"].to_numpy()
+    )
+    heatmap_df.index = [" ++ ".join(index) for index in heatmap_df_index]
+    heatmap_df.index.name = "model1 ++ model2"
+    if "media" in heatmap_df.columns:
+        media_list = heatmap_df["media"].tolist()
+        new_index = [
+            f"{models} in {media_list[i]}" for i, models in enumerate(heatmap_df.index)
+        ]
+        heatmap_df.index = new_index
+        heatmap_df.index.name = "model1 ++ model2 in Media"
+    heatmap_df = heatmap_df.loc[~heatmap_df.index.duplicated(), :]
+    heatmap_df = heatmap_df.drop(["model1", "model2"], axis=1)
+    if "media" in heatmap_df:
+        heatmap_df = heatmap_df.drop(["media"], axis=1)
+    costless = re.compile(r"(?<=\s\()(\d)(?=\))")
+    if "MIP_model1 (costless)" in heatmap_df.columns:
+        mip_model1, mip_model2 = [], []
+        for e in heatmap_df["MIP_model1 (costless)"]:
+            if e == "":
+                mip_model1.append("")
+                continue
+            mip_model1.append(
+                costless.search(str(e)).group() if e not in [0, "0"] else ""
+            )
+        for e in heatmap_df["MIP_model2 (costless)"]:
+            if e == "":
+                mip_model2.append("")
+                continue
+            mip_model2.append(
+                costless.search(str(e)).group() if e not in [0, "0"] else ""
+            )
+        for col, lis in {
+            "c_MIP1": mip_model1,
+            "c_MIP2": mip_model2,
+            "MIP_model1": heatmap_df["MIP_model1 (costless)"].apply(remove_metadata),
+            "MIP_model2": heatmap_df["MIP_model2 (costless)"].apply(remove_metadata),
+        }.items():
+            heatmap_df[col] = to_numeric(lis, errors="coerce")
+    for col in [
+        "MRO_model1",
+        "MRO_model2",
+        "BSS_model1",
+        "BSS_model2",
+        "PC_model1",
+        "PC_model2",
+        "FS",
+        "GYD",
+    ]:
+        if col not in heatmap_df:
+            print(f"The {col} is not computed")
+            continue
+        heatmap_df[col] = to_numeric(
+            heatmap_df[col].apply(remove_metadata), errors="coerce"
+        )
+    del (
+        heatmap_df["BIT"],
+        heatmap_df["MIP_model1 (costless)"],
+        heatmap_df["MIP_model2 (costless)"],
+    )  # TODO colorize the BIT entries as well
+    heatmap_df = heatmap_df.astype(float)
+    int_cols = ["CIP", "MIP_model1", "MIP_model2"]
+    if "costless_MIP_model1" in heatmap_df.columns:
+        int_cols.extend(["c_MIP1", "c_MIP2"])
+    for col in int_cols:
+        heatmap_df[col] = heatmap_df[col].apply(convert_to_int)
+
+    # construct a metabolites table
+    from pandas import DataFrame
+
+    # from pandas import set_option
+    # set_option("display.max_colwidth", None, 'display.width', 1500)
+    ## Process the score metabolites
+    mro_mets, mro_mets_names, mip_model1_mets, mip_model1_mets_names = (
+        [],
+        [],
+        [],
+        [],
+    )
+    mip_model2_mets, mip_model2_mets_names, cip_mets, cip_mets_names = (
+        [],
+        [],
+        [],
+        [],
+    )
+    from json import dump, load
+
+    cpdNames_path = os.path.join(package_dir, "data", "compoundNames.json")
+    with open(cpdNames_path, "r") as jsonIn:
+        cpdNames = load(jsonIn)
+    update_cpdNames = {}
+    for met in mets:
+        # MRO metabolites
+        mro_metIDs = [
+            metID
+            for metID in map(str, met["MRO metabolites"])
+            if metID not in ["None", None]
+        ]
+        mro_mets.append(", ".join(mro_metIDs))
+        names, update_cpdNames = names_updateCPD(mro_metIDs, update_cpdNames)
+        mro_mets_names.append(", ".join(names))
+        # MIP metabolites
+        mip_model1_metIDs = [
+            metID
+            for metID in map(str, met["MIP model1 metabolites"])
+            if metID not in ["None", None]
+        ]
+        mip_model1_mets.append(", ".join(mip_model1_metIDs))
+        names, update_cpdNames = names_updateCPD(mip_model1_metIDs, update_cpdNames)
+        mip_model1_mets_names.append(", ".join(names))
+        ## model2 MIP metabolites
+        mip_model2_metIDs = [
+            metID
+            for metID in map(str, met["MIP model2 metabolites"])
+            if metID not in ["None", None]
+        ]
+        mip_model2_mets.append(", ".join(mip_model2_metIDs))
+        names, update_cpdNames = names_updateCPD(mip_model2_metIDs, update_cpdNames)
+        mip_model2_mets_names.append(", ".join(names))
+        # CIP metabolites
+        cip_metIDs = [
+            metID
+            for metID in map(str, met["CIP metabolites"])
+            if metID not in ["None", None]
+        ]
+        cip_mets.append(", ".join(cip_metIDs))
+        names, update_cpdNames = names_updateCPD(cip_metIDs, update_cpdNames)
+        cip_mets_names.append(", ".join(names))
+    df_content = {
+        "MRO metabolite names": mro_mets_names,
+        "MRO metabolite IDs": mro_mets,
+        "MIP model1 metabolite names": mip_model1_mets_names,
+        "MIP model1 metabolite IDs": mip_model1_mets,
+        "MIP model2 metabolite names": mip_model2_mets_names,
+        "MIP model2 metabolite IDs": mip_model2_mets,
+        "CIP metabolite names": cip_mets_names,
+        "CIP metabolite IDs": cip_mets,
+    }
+    if update_cpdNames != {}:
+        cpdNames.update(update_cpdNames)
+        with open(cpdNames_path, "w") as jsonOut:
+            dump(cpdNames, jsonOut, indent=3)
+    # print(list(map(len, df_content.values())))
+    mets_table = DataFrame(data=df_content)
+    mets_table.index.name = "Community_index"
+
+    # populate the HTML template with the assembled simulation data from the DataFrame -> HTML conversion
+    content = {
+        "table": df.to_html(table_id="main", classes="display"),
+        "mets_table": mets_table.to_html(),
+        "heatmap": heatmap_df.applymap(lambda x: round(x, 3))
+        .style.background_gradient()
+        .to_html(table_id="heat", classes="display"),
+    }
+    env = jinja2.Environment(
+        loader=jinja2.FileSystemLoader(os.path.join(package_dir, "community")),
+        autoescape=jinja2.select_autoescape(["html", "xml"]),
+    )
+    html_report = env.get_template("commscores_template.html").render(content)
+    with open(export_html_path, "w") as out:
+        out.writelines(html_report)
+    return html_report
